+++ src/protobuf/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod convert;
+
 include!(concat!(
     env!("OUT_DIR"),
     "/walfie.gbf.raidfinder.protocol.rs"
